refactor(stella-cli): extract agent/tools.rs test modules before the file hits the 1500 ceiling
agent/tools.rs stood 27 lines from the hard file-size limit with no
grandfather entry — the next ordinary change there would have been blocked
mid-feature. The three inline test modules (tests, benchmark_tests,
diff_baseline_tests) move verbatim to sibling files under agent/tools/,
keeping their module paths and every test name unchanged, the same
extraction that took stella-tools/src/media.rs from 1569 to 491 in #1914.
The production half stays put at 851 lines with real headroom; no baseline
entry is added.

Closes #1915
